server/requirements.txt
openenv-core>=0.2.1
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.9
